refactor(core): organize and unify feature and property tests
Split FeaturesTest.kt into smaller, focused test files for better maintainability. Unified PropertyFactoryTest.kt into a single data-driven test file to reduce code duplication.
